Added edit text and correct callbacks on View Idea Screen Renamed Create Item to View Idea
