+++ app/src/main/java/uk/co/markthomasstevenson/ideame/views/viewidea/FunctionalityAdapter.kt
