+++ app/src/main/res/layout/fragment_view_idea.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".views.createidea.CreateIdeaFragment">
+    tools:context=".views.viewidea.ViewIdeaFragment">
     <RelativeLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
