+++ app/src/main/res/navigation/nav_graph.xml
@@ -14,13 +14,28 @@
                 app:destination="@id/createIdeaFragment"
                 app:enterAnim="@anim/nav_default_enter_anim"
                 app:exitAnim="@anim/nav_default_exit_anim"/>
+        <action android:id="@+id/action_ideaListFragment_to_editIdeaFragment"
+                app:destination="@id/editIdeaFragment"
+                app:enterAnim="@anim/nav_default_enter_anim"
+                app:exitAnim="@anim/nav_default_exit_anim"/>
     </fragment>
     <fragment android:id="@+id/createIdeaFragment"
-              android:name="uk.co.markthomasstevenson.ideame.views.createidea.CreateIdeaFragment"
+              android:name="uk.co.markthomasstevenson.ideame.views.viewidea.ViewIdeaFragment"
               android:label="@string/idea_creation_title"
-              tools:layout="@layout/fragment_create_idea">
+              tools:layout="@layout/fragment_view_idea">
         <action
             android:id="@+id/action_createIdeaFragment_to_ideaListFragment"
+            app:destination="@id/ideaListFragment"
+            app:popUpTo="@id/ideaListFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+
+    <fragment android:id="@+id/editIdeaFragment"
+        android:name="uk.co.markthomasstevenson.ideame.views.viewidea.ViewIdeaFragment"
+        android:label="@string/idea_edit_title"
+        tools:layout="@layout/fragment_view_idea">
+        <action
+            android:id="@+id/action_editIdeaFragment_to_ideaListFragment"
             app:destination="@id/ideaListFragment"
             app:popUpTo="@id/ideaListFragment"
             app:popUpToInclusive="true"/>
